backend/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
groq
python-dotenv
python-multipart
